Move LicenseCard
