+++ site/src/pages/DeploySettingsPage/LicensesSettingsPage/LicensesSettingsPageView.tsx
@@ -5,7 +5,7 @@ import AddIcon from "@mui/icons-material/AddOutlined"
 import RefreshIcon from "@mui/icons-material/Refresh"
 import { GetLicensesResponse } from "api/api"
 import { Header } from "components/DeploySettingsLayout/Header"
-import { LicenseCard } from "components/LicenseCard/LicenseCard"
+import { LicenseCard } from "./LicenseCard"
 import { Stack } from "components/Stack/Stack"
 import { FC } from "react"
 import Confetti from "react-confetti"
